% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axytrek_to_Arrow.R
\name{axytrek_acc_to_dataset}
\alias{axytrek_acc_to_dataset}
\title{Reads axytrek accelerometer data and adds it to an arrow dataset}
\usage{
axytrek_acc_to_dataset(
  files,
  deployments,
  output_dataset,
  date_format = "\%Y-\%m-\%d \%H:\%M:\%OS",
  timezone = "UTC"
)
}
\arguments{
\item{files}{List of axytrek .csv files, file names must include the dep_id in deployments}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}

\item{date_format}{Date format used in .csv files, default is '%Y-%m-%d %H:%M:%OS'}

\item{timezone}{Timezone of raw data, default is UTC}
}
\value{
return Code checks that data meet formatting requirements. Produces a plot
showing a time series of x-axis data from the accelerometer. Deployed data are shown in black,
undeployed data are shown in red, and deployment start/end times are shown with
dashed vertical blue lines.

Output is saved to an arrow dataset at the location indicated by output_dataset.
}
\description{
Reads axytrek accelerometer data and adds it to an arrow dataset
}
