% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axytrek_to_Arrow.R
\name{axytrek_tdr_check}
\alias{axytrek_tdr_check}
\title{Runs checks on axytrek tdr data}
\usage{
axytrek_tdr_check(
  file,
  date_format = "\%Y-\%m-\%d \%H:\%M:\%OS",
  timezone = "UTC"
)
}
\arguments{
\item{date_format}{Date format used in .csv files, default is '%Y-%m-%d %H:%M:%OS'}

\item{timezone}{Timezone of raw data, default is UTC}

\item{files}{List of axytrek .csv files, file names must include the dep_id in deployments}
}
\value{
Ends dataset importing if file names or date formats do not meet checks
}
\description{
Runs checks on axytrek tdr data
}
\examples{

}
