% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Axytrek_to_Arrow.R
\name{axytrek_to_dataset}
\alias{axytrek_to_dataset}
\title{Convert Axytrek biologger data to Arrow dataset}
\usage{
axytrek_to_dataset(
  files,
  deployments,
  output_dataset,
  date_format = "\%Y-\%m-\%d \%H:\%M:\%OS",
  timezone = "UTC",
  acc = T,
  tdr = T,
  gps = T
)
}
\arguments{
\item{files}{List of Axytrek .csv files, data must have file names taht are a partial match to the dep_id in deployment data}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}

\item{date_format}{The POSIXct date format used in .csv files}

\item{timezone}{Timezone of raw data, default is UTC}

\item{plot}{Should plot of data be generated, TRUE or FALSE?}
}
\description{
Reads Axytrek biologger data and parses it into three arrow datasets for ACC, GPS, and TDR data
}
