% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cattrack_to_Arrow.R
\name{cattrack_gps_to_dataset}
\alias{cattrack_gps_to_dataset}
\title{Convert Cattrack GPS data to Arrow dataset}
\usage{
cattrack_gps_to_dataset(data, deployments, output_dataset, plot = T)
}
\arguments{
\item{data}{Data passed from seabiRds::cattrack_to_dataset()}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}

\item{plot}{Should plot of data be generated, TRUE or FALSE?}
}
\value{
Code checks that data meet formatting requirements. Produces a plot
showing a time series of distance from the colony and a map of locations from the GPS.
Deployed data are shown in black, undeployed data are shown in red, and
deployment start/end times are shown with dashed vertical red lines.

Output is saved to an arrow dataset at the location indicated by output_dataset.
}
\description{
Reads Cattrack GPS data and adds it to an arrow dataset
}
