% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cattrack_to_Arrow.R
\name{cattrack_to_dataset}
\alias{cattrack_to_dataset}
\title{Convert Cattrack biologger data to Arrow dataset}
\usage{
cattrack_to_dataset(
  files,
  deployments,
  output_dataset,
  timezone = "UTC",
  plot = T
)
}
\arguments{
\item{files}{List of Cattrack .csv files, data must include logger IDs that match the gps_id in deployment data}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}

\item{timezone}{Timezone of raw data, default is UTC}

\item{plot}{Should plot of data be generated, TRUE or FALSE?}
}
\description{
Reads Cattrack biologger data and parses it into two arrow datasets for GPS and TDR
}
