% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{check_overlaps}
\alias{check_overlaps}
\title{Check for overlapping deployments for metal band or logger id fields}
\usage{
check_overlaps(
  deployments,
  variables = c("metal_band", "gps_id", "gps_id", "tdr_id", "acc_id", "gls_id",
    "mag_id"),
  verbose = FALSE
)
}
\arguments{
\item{deployments}{data.frame with deployment data formatted using seabiRds::formatDeployments()}

\item{variables}{list of variable names to check for overlapping deployments, options include:
"metal_band", "gps_id", "gps_id","tdr_id","acc_id","gls_id","mag_id"}

\item{verbose}{logical. Print a success message if TRUE}
}
\value{
An error message is returned if overlapping deployments are found for the same band or logger_id
}
\description{
Check for overlapping deployments for metal band or logger id fields
}
