% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{checkAxes}
\alias{checkAxes}
\title{Detect and correct switched X and Y axes}
\usage{
checkAxes(data, fix = F, force = F)
}
\arguments{
\item{data}{Vector of accelerometer X value}

\item{fix}{If TRUE, function will flip the x and y axes if they appear flipped}

\item{force}{If TRUE, function will flip the x and y axes whether or not they appear flipped}
}
\description{
Detect and correct switched X and Y axes
}
\details{
Different models of accelerometers can have the x and y channels flipped. This is particularly
an issue between different configurations of Technosmart AxyTrek units. This functions takes formatted
accelerometer data, calculates pitch and roll, then checks if the standard deviation in pitch is higher
than the standard deviation in roll. If this is the case, then it is likely that the axes are switched. By default
the function will only warn give a warning.

If fix is set to true, then the function will return the original dataframe with the x and y axes re-labelled, so that
the x axis is the surge and the y axis is the sway.
}
