% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_GPS.R
\name{cleanGPSData}
\alias{cleanGPSData}
\title{Cleans up GPS data, by clipping to deployment times and filtering unrealistic locations.}
\usage{
cleanGPSData(data, deployments, speedThreshold = NA, plot = T)
}
\arguments{
\item{data}{Name of object with formatted GPS data.}

\item{deployments}{Name of object with formatted deployment data.}

\item{speedThreshold}{Fastest possible movement in km/hr.}

\item{plot}{Should data be plotted (TRUE or FALSE).}

\item{tagTZ}{Timezone of GPS data.}
}
\value{
A new dataframe containing cleaned GPS data.
}
\description{
Cleans up GPS data, by clipping to deployment times and filtering unrealistic locations.
}
\details{
This function clips the GPS data to the start and end times of each GPS deployment. The link between deployment
times and the GPS data is made using the 'dep_id' field. Because
Ecotone GPS units are remotely downloaded and data from multiple devices are stored in the same output file, this connection
is missing. In this case the function creates a 'dep_id' field in each data frame by concatenating the tag
}
