% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_GPS.R
\name{cleanTDRData}
\alias{cleanTDRData}
\title{Cleans up TDR data, by clipping to deployment times}
\usage{
cleanTDRData(data, deployments, plot = T)
}
\arguments{
\item{data}{Name of object with formatted TDR data.}

\item{deployments}{Name of object with formatted deployment data.}

\item{plot}{Should data be plotted (TRUE or FALSE).}

\item{tagTZ}{Timezone of TDR data.}
}
\value{
A new dataframe containing cleaned TDR data.
}
\description{
Cleans up TDR data, by clipping to deployment times
}
\details{
This function clips the TDR data to the start and end times of each TDR deployment. The link between deployment
times and the TDR data is made using the 'dep_id' field. Because
}
