% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecotone_to_Arrow.R
\name{ecotone_dive_to_dataset}
\alias{ecotone_dive_to_dataset}
\title{Convert Ecotone dive summary data to Arrow dataset}
\usage{
ecotone_dive_to_dataset(data, deployments, output_dataset)
}
\arguments{
\item{data}{Data passed from seabiRds::ecotone_to_dataset()}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}
}
\value{
Code checks that data meet formatting requirements. Dive summary data
(start, end, duration) is extracted from tracking data.

Output is saved to an arrow dataset at the location indicated by output_dataset.
}
\description{
Reads Ecotone dive summary data and adds it to an arrow dataset
}
