% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecotone_to_Arrow.R
\name{ecotone_tdr_to_dataset}
\alias{ecotone_tdr_to_dataset}
\title{Convert Ecotone TDR data to Arrow dataset}
\usage{
ecotone_tdr_to_dataset(data, deployments, output_dataset, plot = T)
}
\arguments{
\item{data}{Data passed from seabiRds::ecotone_to_dataset()}

\item{deployments}{Deployment data formatted using seabiRds::formatDeployments()}

\item{output_dataset}{Path where arrow dataset should be saved}

\item{plot}{Should plot of data be generated, TRUE or FALSE?}
}
\value{
Code checks that data meet formatting requirements. Produces a plot
showing a time series of depth data from the TDR Deployed data are shown in black,
undeployed data are shown in red, and deployment start/end times are shown with
dashed vertical red lines.

Output is saved to an arrow dataset at the location indicated by output_dataset.
}
\description{
Reads Ecotone TDR data and adds it to an arrow dataset
}
