% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{fixAxes}
\alias{fixAxes}
\title{Correct switched X and Y axes}
\usage{
fixAxes(data)
}
\arguments{
\item{data}{A formatted dateframe of accelerometer data, including fields: x, y, z, dep_id}
}
\description{
Correct switched X and Y axes
}
\details{
Different models of accelerometers can have the x and y channels flipped. This is particularly
an issue between different configurations of Technosmart AxyTrek units. This functions takes formatted
accelerometer data, swaps the x and y axes and multiplies the original x-axis by -1. This makes the x, y, and z
fields consistent with other acceleromters.
}
