% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getColDist}
\alias{getColDist}
\title{Calculates the distance between the colony (or any fixed location) and each point in a track.}
\usage{
getColDist(lon, lat, colonyLon, colonyLat)
}
\arguments{
\item{lon}{A vector of latitude values.}

\item{colonyLon}{Longitude of the colony (or fixed location).}

\item{colonyLat}{Latitude of the colony (or fixed location).}
}
\value{
A vector of values (in km) giving the distance between each location in the track and the colony (or fixed location).
}
\description{
Calculates the distance between the colony (or any fixed location) and each point in a track.
}
