% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{getDBA}
\alias{getDBA}
\title{Calculate dynamic body acceleration}
\usage{
getDBA(X, Y, Z = NULL, time, window, frequency = NULL, partial = F)
}
\arguments{
\item{X}{Vector of accelerometer X value}

\item{Y}{Vector of accelerometer Y value}

\item{Z}{Vector of accelerometer Z value}

\item{time}{Vector of POSIXct times, should use \%H:\%M:\%OS if sampling frequency >1 Hz}

\item{frequency}{Sampling frequency of data, in Hz, if missing the function will estimate frequency from time}

\item{partial}{If TRUE, calculates partial dynamic body acceleratio}
}
\description{
Calculate dynamic body acceleration
}
