% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getDist}
\alias{getDist}
\title{Calculates the distance between the current location and the previous location in a track.}
\usage{
getDist(lon, lat)
}
\arguments{
\item{lon}{A vector of longitude values.}

\item{lat}{A vector of latitude values.}
}
\value{
A vector of values (in km) giving the distance between each location in the track and the previous location.
}
\description{
Calculates the distance between the current location and the previous location in a track.
}
\section{Warning}{

This function assumes that longitude and latitude values are in chronological order and all values are from a single individual.
Make sure that your data are ordered, and use a for loop or some other method to apply this to multiple tracks
}

