% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getDT}
\alias{getDT}
\title{Calculates the time between between the current location and the previous location in a track.}
\usage{
getDT(time, units = "hours")
}
\arguments{
\item{time}{A vector of time values in POSIXct format.}

\item{units}{The units of time for the output (eg. "days","hours","min","sec")}
}
\value{
A numeric vector of values in the requested time unit, giving the time between each location in the track and the previous location.
}
\description{
Calculates the time between between the current location and the previous location in a track.
}
\section{Warning}{

This function assumes that time values are in chronological order and all values are from a single individual.
Make sure that your data are ordered, and use a for loop or some other method to apply this to multiple tracks
}

