% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getFrequency}
\alias{getFrequency}
\title{Calculates dominant sampling frequency from a list of times}
\usage{
getFrequency(time)
}
\description{
Calculates dominant sampling frequency from a list of times
}
