% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getMode}
\alias{getMode}
\title{Find the modal value in a vector}
\usage{
getMode(x)
}
\arguments{
\item{x}{A vector of values.}
}
\value{
The modal value in \code{x}.
}
\description{
Find the modal value in a vector
}
\examples{
x <- c(1,2,1,2,4,5,7,8,4,2,3,4,4,4)
getMode(x)
}
