% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{getPeaks}
\alias{getPeaks}
\title{Find peak frequency}
\usage{
getPeaks(
  data,
  time,
  method = c("pracma", "fft"),
  window,
  frequency = NULL,
  threshold = 0.1,
  sample = 1
)
}
\arguments{
\item{data}{Vector of numeric values for finding peaks}

\item{time}{Vector of POSIXct times}

\item{window}{Size of moving window for estimating peaks, in seconds}

\item{frequency}{Sampling frequency of data, in Hz}

\item{threshold}{Minimum amplitude to be include as a true peak, results may differ with method}

\item{sample}{Only used in fft method, only calculate paeaks at sample interval, in sec, to save processing time}
}
\description{
The fft method uses a Fast Fourier Transform to identify the peak frequency in the input data over a moving
window defined the frequency * interval. The threshold paramter filters out peaks identified in samples with an inter-quartile range less than the threshold value.
This helps elimnate noisy peaks with low amplitudes. This method can be slow for large data sets. Use sample to only calcualte peaks
at a sampling interval (default is 1 sec), this can greatly speed up the algorithm
}
