% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{getPitch}
\alias{getPitch}
\title{Calculate pitch}
\usage{
getPitch(
  X,
  Y,
  Z,
  window,
  time,
  frequency = NULL,
  standVar = NULL,
  standMin = NULL,
  standMax = NULL
)
}
\arguments{
\item{X}{Vector of accelerometer X value}

\item{Y}{Vector of accelerometer Y value}

\item{Z}{Vector of accelerometer Z value}

\item{window}{Size of moving window for estimating peaks, in seconds}

\item{time}{Vector of POSIXct times, should use \%H:\%M:\%OS if sampling frequency >1 Hz}

\item{frequency}{Sampling frequency of data, in Hz, if missing the function will estimate
frequency from time}

\item{standVar}{Variable to use in standardizing pitch values, if desired}

\item{standMin}{Minimum value in standVar for standarization}

\item{standMax}{Maximum value in standVar for standarization}
}
\description{
Calculate pitch
}
