% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/General_Functions.R
\name{getSessions}
\alias{getSessions}
\title{Creates a unique ID for consecutive values in a time series. This is useful for createing a unique id for all dives or trips}
\usage{
getSessions(value, maxSession = Inf)
}
\arguments{
\item{value}{A vector containing the values you want to create IDs for, can be any data type.}

\item{maxSession}{A numeric value indicating the maximum number of repeating values that should be assigned the same ID. Defaults to Inf for no limit.}
}
\value{
a numeric vector
}
\description{
Creates a unique ID for consecutive values in a time series. This is useful for createing a unique id for all dives or trips
}
