% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualize_tracks.R
\name{plotGPStracks}
\alias{plotGPStracks}
\title{Plot GPS tracks}
\usage{
plotGPStracks(
  data,
  groups = "dep_id",
  legend_label = "Deployments",
  use_legend = T
)
}
\arguments{
\item{data}{GPS tracking data in the seabiRds format}

\item{groups}{name of the variable to use for plotting line/point colours}

\item{legend_label}{Title for legend}

\item{use_legend}{Should the legend be plotted (TRUE/FALSE)}
}
\description{
Creates a simple plot of GPS tracks using a grouping variable for plot colours
}
