% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_GPS.R
\name{prepGPSData}
\alias{prepGPSData}
\title{Reads in raw GPS data, clips tracks to deployment times, and calculates distance from colony.}
\usage{
prepGPSData(inputFolder, deployments, tagTZ = "UTC",
  tagType = "Technosmart", speedThreshold = NA, plot = T)
}
\arguments{
\item{inputFolder}{Folder containing all the raw GPS files to be processed.}

\item{deployments}{Name of object with deployment data.}

\item{tagTZ}{Timezone of GPS data.}

\item{tagType}{Type of GPS biologger used, options are "Technosmart".}

\item{plot}{Should a plot be created for each deployment (TRUE or FALSE).}
}
\value{
A new dataframe containing all GPS data.
}
\description{
Reads in raw GPS data, clips tracks to deployment times, and calculates distance from colony.
}
