% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_GPS.R
\name{readGPSData}
\alias{readGPSData}
\title{Reads in raw GPS data and sets up standard field names.}
\usage{
readGPSData(inputFolder, deployments, tagTZ = "UTC", tagType = "Technosmart")
}
\arguments{
\item{inputFolder}{Folder containing all the raw GPS files to be processed.}

\item{deployments}{Name of object with deployment data.}

\item{tagTZ}{Timezone of GPS data.}

\item{tagType}{Type of GPS biologger used, options are "Technosmart".}
}
\value{
A new dataframe containing all GPS data.
}
\description{
Reads in raw GPS data and sets up standard field names.
}
