% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{checkAxes}
\alias{checkAxes}
\title{Detect and correct switched X and Y axes}
\usage{
checkAxes(data, ask = T, force = F)
}
\arguments{
\item{data}{Vector of accelerometer X value}

\item{ask}{If TRUE, function will plot roll and pitch then ask if axes should be flipped}

\item{force}{If TRUE, function will always flip the x and y axes whether or not they appear flipped}
}
\description{
Detect and correct switched X and Y axes
}
\details{
Different models of accelerometers can have the x and y channels flipped. This is particularly
an issue between different configurations of Technosmart AxyTrek units. This functions provides two options for
switching these axes back.

If force is set to true, then the function will automatically flip the axes, this option should only be used
if you know that all of your data have this issue.

If ask is set to true, then the function will plot pitch and roll using the data, and ask the user if they want to
flip the axes. On most seabirds there hould be more variability in the pitch than in the roll. If the
roll plot appears more variable, then the axes should be flipped.
}
