% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_GPS.R
\name{formatDeployments}
\alias{formatDeployments}
\title{Prepares deployment data for use in clipping tracks}
\usage{
formatDeployments(deployments, birdID, gpsID, onTime, offTime,
  dateFormat = "\%Y-\%m-\%d \%H:\%M", colonyLon = NA, colonyLat = NA,
  gpsFile, depTZ, gpsTZ = "UTC", sinceDate = NULL, keep = NULL)
}
\arguments{
\item{deployments}{A data frame with deployment data.}

\item{birdID}{Character string with name of the field containing bird IDs.}

\item{gpsID}{Character string with name of the field containing GPS IDs.}

\item{onTime}{Character string with name of the field containing deployment start time, must be in a POSIXct compatible format.}

\item{offTime}{Character string with name of the field containing deployment end time, must be in a POSIXct compatible format.}

\item{dateFormat}{Character string specifying the POSIX standard format for times.}

\item{colonyLon}{Longitude of colony (or nest or deployment), default is NULL.}

\item{colonyLat}{Latitude of colony (or nest or deployment), default is NULL.}

\item{depTZ}{Timezone of deployment.}

\item{gpsTZ}{Timezone of GPS data.}

\item{sinceDate}{Limit output to data collected after a certain date.}

\item{keep}{List of variable names for other dpeloyment data to keep with output.}
}
\value{
A new dataframe with deployment times in the same timzone as the GPS data and field names that are compatible with other functions in this package.
}
\description{
Prepares deployment data for use in clipping tracks
}
