% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Process_ACC.R
\name{getODBA}
\alias{getODBA}
\title{Calculate dynamic acceleration}
\usage{
getODBA(X, Y, Z, time, window, frequency = NULL)
}
\arguments{
\item{X}{Vector of accelerometer X value}

\item{Y}{Vector of accelerometer Y value}

\item{Z}{Vector of accelerometer Z value}

\item{time}{Vector of POSIXct times, should use \%H:\%M:\%OS if sampling frequency >1 Hz}

\item{frequency}{Sampling frequency of data, in Hz, if missing the function will estimate frequency from time}
}
\description{
Calculate dynamic acceleration
}
